server/requirements.txt
Flask>=2.3
Flask-Cors>=4.0
Flask-Limiter>=3.4
Flask-SQLAlchemy>=3.1
marshmallow>=3.20
python-dotenv>=1.0
psycopg2-binary>=2.9
PyJWT>=2.8
flasgger>=0.9.7
